requirements.txt
aiogram==3.3.0
python-dotenv==1.0.0
mysql-connector-python==8.2.0
cryptography==41.0.7
aiohttp==3.9.1
python-telegram-bot==20.7 